find - add zsdcc z80 and sccz80 8085 builds for CP/M

+++ mygetopt.c
@@ -10,13 +10,13 @@
 // -HELP
 
 char *optlist[7][3] = {
-    { "-N", 2, 1 },	// -name
-    { "-O", 2, 2 },	// -output
-    { "-U", 2, 3 },	// -user
-    { "-D", 2, 4 },	// -drive
-    { "-H", 1, 5 },	// -help
-    { ".", 1, 6 },	// .
-    { "-A", 1, 7 }	// -alluser
+    { "-N", 2, 1 },		// -name
+    { "-O", 2, 2 },		// -output
+    { "-U", 2, 3 },		// -user
+    { "-D", 2, 4 },		// -drive
+    { "-H", 1, 5 },		// -help
+    { ".", 1, 6 },		// .
+    { "-A", 1, 7 }		// -alluser
 };
 
 #define LENLIST 7
@@ -39,8 +39,8 @@ int mygetopt(int argc, char *argv[], char *list[LENLIST][3])
 	}			// so it must be a '-'
 	if (strstr(argv[getoptindex], optlist[i][0])) {
 	    optarg = argv[getoptindex + 1];
-	    getoptindex += optlist[i][1];
-	    return (optlist[i][2]);
+	    getoptindex += (int) optlist[i][1];
+	    return ((int) optlist[i][2]);
 	}
     }
     return (-1);
